Feature: test jooq generate code with dao

+++ hecate/hecate-persistence/src/main/resources/db/migration/V1__create_global_lock_table.sql
@@ -4,7 +4,7 @@ CREATE TABLE if not exists `global_lock`
     `id`                  int(11)      NOT NULL AUTO_INCREMENT COMMENT '自增 ID',
     `key`                 varchar(255) NOT NULL DEFAULT '' COMMENT '锁',
     `holder`           varchar(255) NOT NULL DEFAULT '' COMMENT '持有者',
-    `expire_ms`            int(11)     NOT NULL DEFAULT -1 COMMENT '锁失效时间，单位为毫秒，默认 -1 代表永久有效',
+    `expire_ms`            bigint(11)     NOT NULL DEFAULT -1 COMMENT '锁失效时间，单位为毫秒，默认 -1 代表永久有效',
     `created_at`          TIMESTAMP    NOT NULL DEFAULT CURRENT_TIMESTAMP COMMENT '创建时间',
     `updated_at`          TIMESTAMP    NOT NULL DEFAULT CURRENT_TIMESTAMP ON UPDATE CURRENT_TIMESTAMP COMMENT '最后修改时间',
     PRIMARY KEY (`id`),
